Add reload feature to create form
